Fix channel list on tablets when team sidebar is present
