+++ app/components/channel_list/index.test.tsx
@@ -8,7 +8,7 @@ import {TeamModel} from '@database/models/server';
 import {renderWithEverything} from '@test/intl-test-helper';
 import TestHelper from '@test/test_helper';
 
-import ChannelsList from './index';
+import ChannelsList from './channel_list';
 
 describe('components/channel_list', () => {
     let database: Database;
@@ -26,7 +26,10 @@ describe('components/channel_list', () => {
 
     it('should match snapshot', () => {
         const wrapper = renderWithEverything(
-            <ChannelsList/>,
+            <ChannelsList
+                isTablet={false}
+                teamsCount={1}
+            />,
             {database},
         );
         expect(wrapper.toJSON()).toMatchSnapshot();
